Move main.py code to pac_bio.py
