+++ bin/generate_pac_bio_id
@@ -19,7 +19,7 @@
 # @author Michael Kubiak <mk35@sanger.ac.uk>
 
 import argparse
-from npg_id_generation import PacBioEntity
+from npg_id_generation.pac_bio import PacBioEntity
 
 parser = argparse.ArgumentParser(
     description="A script to generate a product id for a pac bio product from a given run and well"
